refactor: split text direction CSS evaluation
